refactor: integrate txt parser into pdf module
